text_agent/requirements.txt
streamlit
openai==1.51.2
streamlit-extras
python-dotenv 
plotly
scipy
scikit-learn
azure-search-documents==11.4.0
faiss-cpu
pyodbc==4.0.35
SQLAlchemy==1.4.47
python-dateutil